Fix _fzf_git_files not showing anything in clean repo on macOS
On both Linux and macOS, `grep -f` will match nothing when provided with
an empty file. However, unlike on Linux, when combined with `-v`, grep
on macOS will not match everything when provided with an empty file
while the `-v` flag is active. This results in `_fzf_git_files` being
empty when in a clean repository.

This fixes this behaviour by bypassing the grep command if there is
nothing to exclude.