Fix _fzf_git_files not showing anything in clean repo on macOS (#9)
On both Linux and macOS, `grep -f` will match nothing when provided with
an empty file. However, unlike on Linux, when combined with `-v`, grep
on macOS 10 will not match everything when provided with an empty file
while the `-v` flag is active. This results in `_fzf_git_files` being
empty when in a clean repository.

This fixes this behavior by making the file non-empty by printing `:`.

Co-authored-by: Junegunn Choi <junegunn.c@gmail.com>